Added function to get bobcat camera temperature. Changed SHK to double tap 8bit. Working on getting it to run faster.

+++ src/controller.h
@@ -661,7 +661,7 @@ typedef struct sci_struct{
 } sci_t;
 
 typedef struct shk_struct{
-  uint16 data[SHKXS][SHKYS];
+  uint8 data[SHKXS][SHKYS];
 } shk_t;
 
 typedef struct lyt_struct{
